requirements.txt
# Django Core
Django==4.2.7
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9
django-redis==5.4.0

# PostGIS
GDAL==3.7.3

# REST Framework
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# CORS
django-cors-headers==4.3.1

# Password Hashing
bcrypt==4.1.1

# Rate Limiting
django-ratelimit==4.1.0

# File Storage
boto3==1.29.7
django-storages==1.14.2

# Excel and PDF Export
openpyxl==3.1.2
reportlab==4.0.7
Pillow==10.1.0

# Celery
celery==5.3.4
redis==5.0.1

# Fedapay
fedapay==0.1.8

# SMS Gateway (generic requests library)
requests==2.31.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
hypothesis==6.92.1

# Code Quality
flake8==6.1.0
black==23.11.0
